[guest-bin] move panic handler to guest-bin lib
