+++ src/hyperlight_guest_capi/src/lib.rs
@@ -3,6 +3,10 @@
 
 extern crate alloc;
 
+// TODO(danbugs): this is needed so the panic handler is actually brought in.
+// We can remove it later once more functionality was moved to hyperlight-guest-bin
+extern crate hyperlight_guest_bin;
+
 pub mod dispatch;
 pub mod error;
 pub mod flatbuffer;
