+++ src/tests/rust_guests/simpleguest/Cargo.toml
@@ -5,5 +5,6 @@ edition = "2021"
 
 [dependencies]
 hyperlight-guest = { path = "../../../hyperlight_guest" }
+hyperlight-guest-bin = { path = "../../../hyperlight_guest_bin" }
 hyperlight-common = { path = "../../../hyperlight_common", default-features = false }
 log = {version = "0.4", default-features = false }
